+++ Assets/Scripts/Authoring/PlayerWeaponAuthoring.cs
@@ -14,7 +14,7 @@ public class PlayerWeaponAuthoring : MonoBehaviour
 
 public struct PlayerWeapon : IComponentData {
     [GhostField] public float firerate; // Cooldown in seconds
-    [GhostField] public int damage;
+    [GhostField] public float damage; // Changed from int to float
     [GhostField] public float bulletSpeed;
     [GhostField] public int magSize; // Not used yet, but will be for reloading
 }
